extrator/requirements.txt
fastapi
uvicorn

sqlalchemy
psycopg2-binary
pydantic

python-dotenv
httpx

pdftotext
prometheus-fastapi-instrumentator